% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-preprocessing.R
\name{add_goal_order.restimote}
\alias{add_goal_order.restimote}
\title{Adds goal order vector to determine in what order goals came}
\usage{
\method{add_goal_order}{restimote}(obj, order)
}
\arguments{
\item{obj}{Restimote object}

\item{order}{vector of goal order. eg. (1, 3, 5, 1)}
}
\value{
RestimoteObject witha added field
}
\description{
Adds goal order vector to determine in what order goals came
}
