% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-preprocessing.R
\name{add_goal_positions}
\alias{add_goal_positions}
\title{Adds goal positions to the restimote object}
\usage{
add_goal_positions(obj, positions)
}
\arguments{
\item{obj}{Restimote object.}

\item{positions}{Data frame with goal positions. eg. Needs to have name, position.x and position.y columns}
}
\value{
RestimoteObject with goal positions field
}
\description{
Adds goal positions to the restimote object
}
