% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-loading.R
\name{add_location}
\alias{add_location}
\title{Add locations}
\usage{
add_location(obj, name, x, y)
}
\arguments{
\item{obj}{RestimoteObject}

\item{name}{name of hte location}

\item{x}{x position}

\item{y}{y position}
}
\description{
Add locations
}
