% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-preprocessing.R
\name{calibrate_compass}
\alias{calibrate_compass}
\title{Fixes possible compass offsets created by iPhone irregular comapss settings}
\usage{
calibrate_compass(obj, estimote_offset)
}
\arguments{
\item{obj}{RestimoteObject}

\item{estimote_offset}{Offset of the map as set in the estimote Layout}
}
\value{
restimote object
}
\description{
Estimote sets each map to a certain orientation and then substracts that from compass information it is getting. 
But it happens, that iPhone compass calibrates like crazy. This function searches for parts of the log that are 
CAlibrated using the calibrate function in the restimote app and then separately preprocesses those parts where 
researcher asked participant to calibrate manually
}
\examples{
obj <- calibrate_compass(obj, 321)

}
