% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{create_plot}
\alias{create_plot}
\title{Create a plot wiht constraints}
\usage{
create_plot(obj)
}
\arguments{
\item{obj}{RestimoteObject}
}
\value{
ggplot2 plot
}
\description{
Create a plot wiht constraints
}
