% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-cleaning.R
\name{find_true_trial_log}
\alias{find_true_trial_log}
\title{finding the place that is the closest to the starting position ..... it might come sooner 
than person really started this trial, but it doesn't matter, as times are calculated from a 
differnet log - we need this for the distance summary}
\usage{
find_true_trial_log(obj, trialId, benevolence = 30, radius = 2)
}
\description{
finding the place that is the closest to the starting position ..... it might come sooner 
than person really started this trial, but it doesn't matter, as times are calculated from a 
differnet log - we need this for the distance summary
}
