% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_action_times.restimote}
\alias{get_action_times.restimote}
\title{Returns times when certain action occured during entire experiment}
\usage{
get_action_times.restimote(obj, action)
}
\arguments{
\item{obj}{Restimote object with companion loaded}

\item{action}{string with the name of the action}

\item{trialId}{integer designating which trial to load}
}
\value{
vector of times
}
\description{
Returns times when certain action occured during entire experiment
}
