% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_companion}
\alias{get_companion}
\title{Returns companion data frame with all the actions experiemnter logged}
\usage{
get_companion(obj)
}
\arguments{
\item{obj}{}
}
\value{
data.frame with the log
}
\description{
Returns companion data frame with all the actions experiemnter logged
}
