% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_companion_timewindow}
\alias{get_companion_timewindow}
\title{Returns companion from particular time to another time}
\usage{
get_companion_timewindow(obj, start, end)
}
\arguments{
\item{obj}{Restimote object with companion loaded}

\item{start}{Starting time of the log}

\item{end}{last accepable time of the log}
}
\value{
data.frame companion log
}
\description{
Returns companion from particular time to another time
}
