% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_goal_position}
\alias{get_goal_position}
\title{returns 2D vector with x and Y position of trial goal position}
\usage{
get_goal_position(obj, trialId)
}
\arguments{
\item{obj}{Restimote object}

\item{trialId}{what trial}
}
\value{
2D vector with x and Y of the goal position
}
\description{
returns 2D vector with x and Y position of trial goal position
}
