% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_log.restimote}
\alias{get_log.restimote}
\title{Returns log object with positions during the experiment}
\usage{
\method{get_log}{restimote}(obj)
}
\arguments{
\item{obj}{}
}
\value{
data.frame with the log
}
\description{
Returns log object with positions during the experiment
}
