% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_log_timewindow.restimote}
\alias{get_log_timewindow.restimote}
\title{Returns log from particular time to another time}
\usage{
\method{get_log_timewindow}{restimote}(obj, start, end)
}
\arguments{
\item{obj}{Restimote object with log loaded}

\item{start}{Starting time of the log}

\item{end}{last accepable time of the log}
}
\value{
data.frame participant log
}
\description{
Returns log from particular time to another time
}
