% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_n_actions.restimote}
\alias{get_n_actions.restimote}
\title{Returns how many actions of particular type were recorded}
\usage{
\method{get_n_actions}{restimote}(obj, actions)
}
\arguments{
\item{action}{string with searched action name}

\item{...}{extra parameters}
}
\value{
integer with number of recorded actions
}
\description{
Returns how many actions of particular type were recorded
}
