% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-restimote.R
\name{get_n_pointings}
\alias{get_n_pointings}
\title{Gets number of times participant should point and pointed}
\usage{
get_n_pointings(obj)
}
\arguments{
\item{obj}{RestimoteObject}
}
\value{
list with log, companion fields
}
\description{
Gets number of times participant should point and pointed
}
