% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters-restimote.R
\name{get_position_trial}
\alias{get_position_trial}
\title{Gets position data inside trial}
\usage{
get_position_trial(obj, trialId)
}
\arguments{
\item{obj}{RestimoteObject object}

\item{trialID}{int designating which trial to select}
}
\value{
log data.frame with only rows from during the trial
}
\description{
Gets position data inside trial
}
