% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_start_position}
\alias{get_start_position}
\title{returns vector 2 of x and Y position of trial goal position}
\usage{
get_start_position(obj, trialId)
}
\arguments{
\item{obj}{}

\item{trialId}{}
}
\value{
vector 2 of x and Y of the starting position
}
\description{
returns vector 2 of x and Y position of trial goal position
}
