% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_action_times.restimote}
\alias{get_trial_action_times.restimote}
\title{Returns times when certain action occured during a trial}
\usage{
\method{get_trial_action_times}{restimote}(obj, trialId, action)
}
\arguments{
\item{obj}{Restimote object with companion loaded}

\item{trialId}{integer designating which trial to load}

\item{action}{string with the name of the action}
}
\value{
vector of times
}
\description{
Returns times when certain action occured during a trial
}
