% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_distance.restimote}
\alias{get_trial_distance.restimote}
\title{Returns distance traveled during a particular trial}
\usage{
\method{get_trial_distance}{restimote}(obj, trialId, true_log = T,
  benevolence = 20, radius = 5)
}
\arguments{
\item{obj}{}

\item{trialId}{}

\item{true_log}{}

\item{benevolence}{only valid if true_log. See benevolence in true_trial_log}

\item{radius}{only valid if true_log. See radius in true_trial_log}
}
\description{
Returns distance traveled during a particular trial
}
