% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_log.restimote}
\alias{get_trial_log.restimote}
\title{Gets position data inside trial}
\usage{
\method{get_trial_log}{restimote}(obj, trialId)
}
\arguments{
\item{obj}{RestimoteObject object}

\item{trialID}{int designating which trial to select}
}
\value{
log data.frame with only rows from during the trial
}
\description{
Gets position data inside trial
}
