% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_point_orientation.restimote}
\alias{get_trial_point_orientation.restimote}
\title{Gets the orientation of pointing in particular trial}
\usage{
get_trial_point_orientation.restimote(obj, trialId)
}
\arguments{
\item{obj}{RestimoteObject}

\item{trialId}{integer with valid trial id}
}
\value{
orientation fo the participant when they pointed
}
\description{
Gets the orientation of pointing in particular trial
}
