% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_point_times.restimote}
\alias{get_trial_point_times.restimote}
\title{Returns of times when people were pointing}
\usage{
get_trial_point_times.restimote(obj, pointId, viewpoint = NULL)
}
\arguments{
\item{pointId}{which point you want}

\item{viewpoint}{if set, only returns pointings from given viewpoint}
}
\value{
list with start and end. for last trial, end is NA, as there is no finishing signal
}
\description{
Returns of times when people were pointing
}
