% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_times.restimoter}
\alias{get_trial_times.restimoter}
\title{Colelcts start and end times for trial of particular id}
\usage{
\method{get_trial_times}{restimoter}(obj, trialId)
}
\arguments{
\item{obj}{Restimote object}

\item{trialId}{}
}
\value{
List with start and end. For the last trial,
}
\description{
Colelcts start and end times for trial of particular id
}
