% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-restimote-companion-log.R
\name{load_restimote_companion_log}
\alias{load_restimote_companion_log}
\title{Opens restimote companion app in a given folder}
\usage{
load_restimote_companion_log(filepath, timestamp = NULL, obj = NULL)
}
\arguments{
\item{filepath}{either a filepath, or directory where the file is located}

\item{timestamp}{in case multiple restimotes are in the directory}

\item{obj}{already created restimote object}
}
\value{

}
\description{
Opens restimote companion app in a given folder
}
