% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_add_restimote_path}
\alias{plot_add_restimote_path}
\title{plot entire path log from the restimote}
\usage{
plot_add_restimote_path(plt, df_pos)
}
\arguments{
\item{obj}{}
}
\value{
plot with walked path
}
\description{
plot entire path log from the restimote
}
