% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_add_trial_start_goal}
\alias{plot_add_trial_start_goal}
\title{Title}
\usage{
plot_add_trial_start_goal(plt, obj, trialId)
}
\arguments{
\item{plt}{plt to have the start and goal added to}

\item{obj}{Restimote object with goal positions and goal order}

\item{trialId}{valid trialID}
}
\value{
plot
}
\description{
Title
}
