% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_path_time}
\alias{plot_path_time}
\title{Plots path between two timepoints}
\usage{
plot_path_time(obj, start, end)
}
\arguments{
\item{obj}{RestimoteObject}

\item{start}{time when we want to start plotting}

\item{end}{time at which we shoudl stop plotting}
}
\value{
plot
}
\description{
Plots path between two timepoints
}
\examples{

}
