% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_trial_path}
\alias{plot_trial_path}
\title{Title}
\usage{
plot_trial_path(obj, trialId)
}
\arguments{
\item{obj}{RestimoteObject Needs to be preprocessed}

\item{trialId}{integed of given trial}
}
\value{
prints plot
}
\description{
Title
}
\examples{
plot_trial_path(obj, 1)
}
