% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_trials_paths}
\alias{plot_trials_paths}
\title{Title}
\usage{
plot_trials_paths(obj, columns = 5, indices = c())
}
\arguments{
\item{obj}{}

\item{columns}{}

\item{indices}{}
}
\description{
Title
}
