% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-visualising.R
\name{plot_true_trial_path}
\alias{plot_true_trial_path}
\title{Plots path as calculated by the true trial log, rather than by given timestamps}
\usage{
plot_true_trial_path(obj, trialId)
}
\arguments{
\item{obj}{}

\item{trialId}{}
}
\description{
Plots path as calculated by the true trial log, rather than by given timestamps
}
