% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-restimote.R
\name{preprocess_companion_log}
\alias{preprocess_companion_log}
\title{Preprocessses and returns RestimoteObject companion}
\usage{
preprocess_companion_log(obj)
}
\arguments{
\item{obj}{RestimoteObject}
}
\value{
preprocesse RestimoteObject
}
\description{
Preprocessses and returns RestimoteObject companion
}
