% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-restimote.R
\name{remove_companion_ids}
\alias{remove_companion_ids}
\title{Title}
\usage{
remove_companion_ids(obj, action, ids)
}
\arguments{
\item{obj}{RestimoteObject. Needs to be preprocessed}

\item{action}{Name of the indexed event that you want ot remove}

\item{ids}{vector of Ids to be removed}
}
\value{
RestimoteObject with removed Ids and recomputed new ids
}
\description{
Title
}
