% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-restimote.R
\name{remove_pointing}
\alias{remove_pointing}
\title{Remove pointing mistakes from companion}
\usage{
remove_pointing(obj, ids)
}
\arguments{
\item{obj}{RestimoteObject. Needs to be preprocessed}

\item{ids}{vector of Ids to be removed}
}
\value{
RestimoteObject with removed Ids and recomputed new ids
}
\description{
Remove pointing mistakes from companion
}
