% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-cleaning.R
\name{remove_random_points}
\alias{remove_random_points}
\title{Title}
\usage{
remove_random_points(df_player, allowed_speed = 4)
}
\arguments{
\item{df_player}{data fram withg player position}

\item{allowed_speed}{what is the maximum allowed sppeed before it is assumed to be artifact}
}
\value{
smoothed df_player
}
\description{
Title
}
