% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-preprocessing.R
\name{remove_trials}
\alias{remove_trials}
\title{Remove trial mistakes from companion}
\usage{
remove_trials(obj, ids)
}
\arguments{
\item{obj}{RestimoteObject. Needs to be preprocessed}

\item{ids}{vector of Ids to be removed}
}
\value{
RestimoteObject with removed Ids and recomputed new ids
}
\description{
Remove trial mistakes from companion
}
\examples{
obj <- remove_trials(obj, c(1, 4))
obj <- remove_trials(obj, 9)
}
