% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-object.R
\name{RestimoteObject}
\alias{RestimoteObject}
\title{Creates list with parameters to be filled while processng}
\usage{
RestimoteObject()
}
\value{
list with instantiated fields
}
\description{
Creates list with parameters to be filled while processng
}
