% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-cleaning.R
\name{smooth_positions}
\alias{smooth_positions}
\title{Title}
\usage{
smooth_positions(log, type = "median", points = 11)
}
\arguments{
\item{log}{Data framew with Position.X and Position.Y column}

\item{type}{median, approx, spline}

\item{points}{necessary if choosing median smoothing}
}
\value{
log with smoothed X and Y columns
}
\description{
Title
}
