% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-cleaning.R
\name{true_trial_log}
\alias{true_trial_log}
\title{Goes through the restimote log and defines trials not in respect to the 
companion log but in respect to when player achieved started at the start 
and when they reached given goal - times are incorrect, but path shoudl correspond}
\usage{
true_trial_log(obj, trialId, benevolence = 30, radius = 2)
}
\arguments{
\item{obj}{Preprocessed Restimote object}

\item{trialId}{}

\item{benevolence}{how much time after the trial ended you give the function to search (default 30s)}

\item{radius}{what radius around goal position do you consider arriving at the goal position? (defualt 2 meters)}
}
\value{
log with the trial from the beginning to the
}
\description{
Goes through the restimote log and defines trials not in respect to the 
companion log but in respect to when player achieved started at the start 
and when they reached given goal - times are incorrect, but path shoudl correspond
}
