% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_start_position}
\alias{get_start_position}
\title{returns 2D vector with x and Y position of trial start position}
\usage{
get_start_position(obj, trialId)
}
\arguments{
\item{obj}{Restimote object}

\item{trialId}{integer of the trial to fetch}
}
\value{
2D vector with x and Y of the starting position
}
\description{
returns 2D vector with x and Y position of trial start position
}
