% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-getters.R
\name{get_trial_duration.restimote}
\alias{get_trial_duration.restimote}
\title{Returns length of the trial withouth pauses/app resets included}
\usage{
\method{get_trial_duration}{restimote}(obj, trialId, withoutPauses = T)
}
\arguments{
\item{obj}{Restimote object}

\item{trialId}{}

\item{withoutPauses}{bool. Defaults to true}
}
\value{
numeric of time length
}
\description{
Returns length of the trial withouth pauses/app resets included
}
