% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-preprocessing.R
\name{preprocess_restimote_log}
\alias{preprocess_restimote_log}
\title{Preprocesses and returns restimote log}
\usage{
preprocess_restimote_log(obj)
}
\arguments{
\item{obj}{}
}
\value{
preprocessed RestimoteObject
}
\description{
Preprocesses and returns restimote log
}
