% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restimote-cleaning.R
\name{true_trial_log}
\alias{true_trial_log}
\title{Goes through the restimote log and defines trials not in respect to the 
companion log but in respect to when player achieved started at the start 
and when they reached given goal - times are incorrect, but path shoudl correspond}
\usage{
true_trial_log(obj, trialId)
}
\arguments{
\item{obj}{Preprocessed Restimote object}

\item{trialID}{valid id of the trial}
}
\description{
Goes through the restimote log and defines trials not in respect to the 
companion log but in respect to when player achieved started at the start 
and when they reached given goal - times are incorrect, but path shoudl correspond
}
